Add example graph; test with client and graph
