+++ Cargo.toml
@@ -21,6 +21,7 @@ sha2 = "0.10.8"
 tokio = { version = "1.37.0", features = ["full"] }
 esplora-client = { path = "../rust-esplora-client/" }
 serde_json = "1.0.116"
+lazy_static = "1.4.0"
 
 [dev-dependencies]
 rand_chacha = "0.3.1"
